Set a lower PBKDF count for some of the unit tests
The default tuf.keystore._PBKDF_ITERATIONS (90,510) slow down the unit tests considerably.  Reduce it temporarily for the unit tests depending on it and that do not test the strength of derived keys.  All 164 tests now run in approximately 2 minutes, down from approximately 17 minutes.  The lowered PBKDF count set for the unit tests is equal to the previous key derivation count used by evpy.