Updated the developer tool to inherit the Targets object The motivation behind this change is code reutilization and quick updating. Most of the methods called in the old version of the developer tool immediate to the targets counterpart. The calling convention now resembles more objective intended without losing any functionality and with automatic updates from any changes in the repository_tool.
The project object used to have a Targets object contained. This did
make more sense in a conceptual way (a project could have had more roles
than only targets). However, after reviewing the functionality of the
tool, it seems clear that the best way to adopt the changes
automatically is to inherit the targets object.

Soft wrappers are still to be made, we need to override the
add_verification_key method to in order to avoid having more than one
key on the project.