Allow partial metadata to be properly loaded with zero good signatures.
Previously, the repository tool allowed partial metadata to be written with zero good signatues and later successfully loaded.  However, the partial metadata was not properly marked by load_repository().  The 'partial_loaded' flag is now set (if partial metadata is written with zero good signatures) and matches the behavior of partial_write().