Test: improve failure behavior of test_slow_retrieval_attack
and also clarify setup workflow in test_slow_retrieval_attack.

Because the test was written incorrectly, the test was failing
with a bad hash error after 31 seconds, when it should instead
have been failing because the slow retrieval attack was not
averted.

In particular, a target file was updated, but metadata was not
correctly updated on the repository and, further, the client's
copy of the metadata was never updated. (The client continued
to expect 31 bytes of target file instead of 400 or 800.)

The way the test used to run, the target file change
previously had no purpose.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>