ASN.1: further expand layers of definitions.... (squash me)
The way I'm performing the generic conversion currently requires
that every level of the definitions be a defined class.

I can't do this currently:
  class SomeDefinition(Sequence):
    _fields = [
      ...
      ('compound_field', Sequence, {'_fields': [
        ('a', Integer),
        ('b', Integer)])

Instead I have to define the compound_field entry as its own class
and then do this:
  class SomeDefinition(Sequence):
    _fields = [
      ...
      ('compound_field', CompoundField)]

This is because the recursion passes around datatypes derived
from the objects it receives, and type() of the compound field
object in the first case above is just a bare "Sequence", which
loses all the information about its fields, while in the second
case the field information is preserved. :/

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>