Fix remaining test_roledb tests, reorganize arg tests
Two functions now exist to replace _test_rolename (which was a bit
of a misleading name), and these are now used to perform argument
testing for roledb functions that query a single role or query
information about a delegation from one role to another role.

In addition, tests for roledb.get_delegated_paths were also
updated, duplicating some of the above for reasons explained in
code comments.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>