Testing: Reset spec version if a failed test changed it
A test in test_updater modifies tuf.SPECIFICATION_VERSION
temporarily.  If it fails, it doesn't get to reset the value.

This commit ensures that tuf.SPECIFICATION_VERSION is an
expected value at the start of each test -- specifically, the
value it had when this test module was first imported.

I thought about using a try/except/else structure, but the span
of the test was too long and I didn't want to nest try/except.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>