Add extra installation instruction to contrib doc
Add hint to install `coverage` before using it.

This should be installed via dev-requirements.txt, however it
does not seem to fit in there, because dev-requirements.txt pins
all its dependencies which does not seem to make sense for a
development tool.

Maybe a hierarchy of requirements.txt similar to
https://github.com/in-toto/in-toto/commit/06a28987dc8092afa1060ee51cd3e19a6f324855
could be established.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>