Rename variable
As per PEP 008 for function and variable names it's
recommended that:

"Function names should be lowercase, with words separated
by underscores as necessary to improve readability.
Variable names follow the same convention as function names."

I am renaming this variable, because "filelength" are two
separate words merged and it's not so readable.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>