Add 'delegating_role' to role metadata
Add a new field in internal role metadata for storing the delegating
role. This is needed since all updates on roledb are performed on
the delegated role which has no knowledge about its 'delegator'.

When a delegated role is marked as dirty the 'delegating_role'
information is used to trigger an update chain.

Update the load_repository() function to load the delegations
metadata starting from 'targets' in order to get the 'delegating_role'.

Signed-off-by: Teodora Sechkova <tsechkova@vmware.com>