Fix generating consistent targets metadata
It should be possible to request consistent targets metadata when using
existing fileinfo, because the metadata itself doesn't differ. Consistent
targets, copies of the file with the file's hash prepended to the target's
file name, should exist in the targets directory but are not listed in the
targets metadata.

Throwing an exception here prevents enabling consistent metadata to be
written when using existing fileinfo structures, instead we log that the
consistent copies of the target files are expected to exist in the
repository's targets directory and continue assuming they are present.

Signed-off-by: Joshua Lock <jlock@vmware.com>