Silence expected "file not found" errors
Currently, when we run the unit tests we get a lot of
messages with the content: "code 404, message File not found"
even a successful run.

This is caused because tuf will always try to download a newer version
of root.json, but as expected, at one point it will download
the freshest version of root.json and when it tries to download
a newer one it will fail with this error message.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>