simple-server: Allow address re-use
If a test happens to use the same port as a previous test run
(either by bad luck or hardcoding like TestMultiRepoUpdater) that
happened within a minute, the second run will fail because TCP by
default keeps sockets open for a while.

Avoid this by explicitly saying re-use is fine in this case.

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>