This WIP commit creates a separate keydb for each delegating instance.
This means that all roles delegated to by root will look for keys
from the root metadata, roles delegated to by targets will look for
keys in targets metadata, and so on.

This commit adds this functionality to the keydb and ensures that
these separate keydbs are created and used for storing keys. It
requires a future change to the roledb to store the delegating role
and further testing.

Signed-off-by: marinamoore <mmoore32@calpoly.edu>