Add targets map file to single-repository updater
This commit adds support for TAP 13 to the Updater
class. TAP 13 allows a client to specify which targets
metadata on the repository they would like to use as the
top-level targets metadata. This mechanism allows
repository managers to split a repository into multiple
namespaces and for clients to specify which of these
namespaces to trust.

This commit implements this feature by reading the
targets map file and using the indicated metadata
as top-level targets metadata.

Signed-off-by: marinamoore <mnm678@gmail.com>