This commit fixes snapshot merkle filenames and adds a dictionary for merkle paths.
snapshot merkle filenames should be of the form role-snapshot.json. This commit ensures that they will not have an additional .json in the middle (role.json-snapshot.json).

In addition, this commit updates the snapshot merkle files to use a dictionary for the merkle path to preserve the order of elements.

Signed-off-by: marinamoore <mnm678@gmail.com>