Allow custom prefixes for hashed bin delegations
This commit allows hashed bin delegations (both succinct and otherwise)
to be created with a custom prefix. If the prefix is not provided,
the existing defaults will be used.

As part of this addition, this commit cleans up some of the checks for
succinct hashed bin delegations to check for the existence of the
delegation_hash_prefix_len field rather than using the format of the name to
determine the type of delegation.

Signed-off-by: marinamoore <mnm678@gmail.com>