Don't use snapshot metadata when snapshot merkle trees are used.
This commit ensures that when snapshot merkle trees are used, the
client is not looking for version information in the snapshot
file, but instead waiting to download version information from
the snapshot merkle tree.

Another approach could be to download the snapshot merkle file
early to ensure that it exists before continuing the verification.

Signed-off-by: marinamoore <mnm678@gmail.com>