Use repo_lib helper functions for bin names
Update the repo_lib helper functions to use a prefix
and create a name using a zero-padded count instead of
the prefix range. So the name will be of the format
prefix-count (for example bin.hbd-02).

In addition, this updates the repository tool to use
the updated repo_lib functions to avoid duplicate code.

Signed-off-by: marinamoore <mnm678@gmail.com>