Imports cleanup: Define package imports
The end goal is to get rid of imports style
    import package.module
as that is incompatible with vendoring. The secondary goal is to
preserve the ability to fully namespace things (so to use
tuf.module.something in the code).

This commit defines the basic set of library imports.
 exceptions
 formats
 keydb
 roledb
 settings
 import sig
These will be available under tuf namespace after "import tuf"

There's also a set of client imports
 download
 mirrors
These will be available under client namespace after
"from tuf import client".

These sets do not include
 log (because importing it has side effects)
 repository_tool
 repository_lib
 developer_tool

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>