Create dictionary to store succinct hashed bin delegation metadata.
This commit creates a 'succinct_hash_delegations' field that contains
'delegation_hash_prefix_len' and 'bin_name_prefix' to be used for
succinct hashed bin delegations. When this field is present, the
'name' field is no longer needed, and so is removed from the metadata.

For more detail, see the succinct hashed bin delegation TAP pr.

Signed-off-by: marinamoore <mnm678@gmail.com>