Make slow_retrieval use utils.py port generation
Until now, slow_retrieval.py couldn't use the TestServerProcess class
port generation and retry mechanism from utils.py because we were
using httpd.handle_request() which handles only ONE request.
Then, what happened is that when we use wait_for_server() to make
a test connection and verify that the server is up,
the slow_retrieval server handles that connection
(which it accepts as a request) and exits.

We avoided that use-case by passing timeout = 0 and avoiding
calling wait_for_server() on this special value.
Now, when we use  httpd.serve_forever() this problem is resolved
and no longer we need to make those checks.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>