tox: Workaround double dep in with-sslib-master
Commit eb00d14 modified requirements-pinned.txt so that sslib specifiers
are now "[crypto,pynacl]". This happens to match the exact specifiers
used for the sslib git master dependency in tox.ini. This triggers pip
to say:
  ERROR: Double requirement given: securesystemslib[crypto,pynacl]==0.16.0
  (from -r /home/jku/src/tuf/requirements-pinned.txt (line 12)) (already
  in securesystemslib[crypto,pynacl] from
  git+http://github.com/secure-systems-lab/securesystemslib.git@master#egg=securesystemslib[crypto,pynacl],
  name='securesystemslib')

Avoid this by not setting any specifiers for the sslib git master
dependency in tox.ini: This makes pip happy and we get the git master
version installed. pynacl and crypto are still installed because they
are in requirements-pinned.txt.

Fixes #1184.

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>