Add tests for tests/utils.py
We added a retry mechanism on server startup failure which
generates a new port and tries to start the server again
until the server is started or the timeout has expired.

We want to make sure that this mechanism is well tested as for
usual cases same for corner cases e.g. when the server hangs on
forever or just exits unexpectedly.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>