Delegate port generation for the tests to the OS
By giving 0 to the port argument we ask the OS to give us
an arbitrary unused port.
This method is much better than if we generate a port, because
even though we are generating a random port there is always
the chance that the port could be already in use.

In this commit, I also make sure that there is no place in the tests
where we are manually generating and passing ports.

Also, because none of the tests pass a port to the TestServerProcess
class and its __init__ function I have removed the "port" argument.

Finally, until now slow_retrieval.py couldn't use the
TestServerProcess class from utils.py port generation because we were
using httpd.handle_request() which handles only ONE request.
Then, what happened was that when we use wait_for_server() to make
a test connection and verify that the server is up,
the slow_retrieval server handles that connection
(which it accepts as a request) and exits.

We avoided that use-case by passing timeout = 0 and avoiding
calling wait_for_server() on this special value.
Now, when we use httpd.serve_forever() this problem is resolved
and no longer we need to make those checks.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>